Api4Ctrl - Fix infinite reloads
I noticed that the `afform-html` GUI is sending an infinite number requests to
reload its data.  After bisecting, I found that e7ebc633a had expanded the
scope the `watch()` so that it reloads on *any* change to `affApi4Ctrl` (not
just a change in params).  But the `affApi4Ctrl` includes the `.result` data
- as soon as the result data is returned, it re-triggers the watch and sends
the request again.  This becomes an infinite loop.

I don't think this was necessary for the goal of indexing - although maybe
the idea was that a change to `.index` should cause a reload, which seems
reasonable.  So I've whitelisted fields which deal with APIv4 inputs.